Split out initialize-specific tests into devices module
Remove any unnecessary use of BlockDevMgr functionality.
