+++ src/engine/strat_engine/backstore/devices.rs
@@ -241,3 +241,169 @@ pub fn wipe_blockdevs(blockdevs: &[StratBlockDev]) -> StratisResult<()> {
         Err(StratisError::Engine(ErrorEnum::Error, err_msg))
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::fs::OpenOptions;
+
+    use rand;
+    use uuid::Uuid;
+
+    use crate::engine::strat_engine::{
+        backstore::metadata::device_identifiers,
+        cmd,
+        tests::{loopbacked, real},
+    };
+
+    use super::*;
+
+    /// Verify that it is impossible to initialize a set of disks of which
+    /// even one of them has a signature.  Choose the dirty disk randomly.
+    fn test_fail_single_signature(paths: &[&Path]) {
+        assert!(paths.len() > 1);
+        let index = rand::random::<u8>() as usize % paths.len();
+
+        cmd::create_fs(paths[index], None).unwrap();
+        cmd::udev_settle().unwrap();
+
+        let pool_uuid = Uuid::new_v4();
+        assert_matches!(
+            initialize(
+                pool_uuid,
+                resolve_devices(paths).unwrap(),
+                MDADataSize::default(),
+                &HashSet::new()
+            ),
+            Err(_)
+        );
+        for (i, path) in paths.iter().enumerate() {
+            if i == index {
+                assert_matches!(
+                    DevOwnership::from_udev_ownership(
+                        &block_device_apply(path, |d| decide_ownership(d))
+                            .unwrap()
+                            .unwrap()
+                            .unwrap(),
+                        path
+                    )
+                    .unwrap(),
+                    DevOwnership::Theirs(_)
+                );
+            } else {
+                assert_matches!(
+                    DevOwnership::from_udev_ownership(
+                        &block_device_apply(path, |d| decide_ownership(d))
+                            .unwrap()
+                            .unwrap()
+                            .unwrap(),
+                        path
+                    )
+                    .unwrap(),
+                    DevOwnership::Unowned
+                );
+            }
+        }
+
+        let clean_paths = paths
+            .iter()
+            .enumerate()
+            .filter(|(n, _)| *n != index)
+            .map(|(_, v)| *v)
+            .collect::<Vec<&Path>>();
+
+        assert_matches!(
+            initialize(
+                pool_uuid,
+                resolve_devices(&clean_paths).unwrap(),
+                MDADataSize::default(),
+                &HashSet::new()
+            ),
+            Ok(_)
+        );
+        cmd::udev_settle().unwrap();
+
+        for path in clean_paths {
+            assert_eq!(
+                pool_uuid,
+                device_identifiers(&mut OpenOptions::new().read(true).open(path).unwrap(),)
+                    .unwrap()
+                    .unwrap()
+                    .0
+            );
+        }
+    }
+
+    #[test]
+    pub fn loop_test_fail_single_signature() {
+        loopbacked::test_with_spec(
+            &loopbacked::DeviceLimits::Range(2, 3, None),
+            test_fail_single_signature,
+        );
+    }
+
+    #[test]
+    pub fn real_test_fail_single_signature() {
+        real::test_with_spec(
+            &real::DeviceLimits::AtLeast(2, None, None),
+            test_fail_single_signature,
+        );
+    }
+
+    #[test]
+    pub fn travis_test_fail_single_signature() {
+        loopbacked::test_with_spec(
+            &loopbacked::DeviceLimits::Range(2, 3, None),
+            test_fail_single_signature,
+        );
+    }
+
+    /// Test that initialing devices claims all and that destroying
+    /// them releases all.
+    fn test_ownership(paths: &[&Path]) {
+        let pool_uuid = Uuid::new_v4();
+        let blockdevs = initialize(
+            pool_uuid,
+            resolve_devices(paths).unwrap(),
+            MDADataSize::default(),
+            &HashSet::new(),
+        )
+        .unwrap();
+
+        cmd::udev_settle().unwrap();
+
+        for path in paths {
+            assert_eq!(
+                pool_uuid,
+                device_identifiers(&mut OpenOptions::new().read(true).open(path).unwrap(),)
+                    .unwrap()
+                    .unwrap()
+                    .0
+            );
+        }
+
+        wipe_blockdevs(&blockdevs).unwrap();
+
+        for path in paths {
+            assert_eq!(
+                device_identifiers(&mut OpenOptions::new().read(true).open(path).unwrap(),)
+                    .unwrap(),
+                None
+            );
+        }
+    }
+
+    #[test]
+    pub fn loop_test_ownership() {
+        loopbacked::test_with_spec(&loopbacked::DeviceLimits::Range(1, 3, None), test_ownership);
+    }
+
+    #[test]
+    pub fn real_test_ownership() {
+        real::test_with_spec(&real::DeviceLimits::AtLeast(1, None, None), test_ownership);
+    }
+
+    #[test]
+    pub fn travis_test_ownership() {
+        loopbacked::test_with_spec(&loopbacked::DeviceLimits::Range(1, 3, None), test_ownership);
+    }
+}
